Improved speed of point-and-select feature and bumped version to 6.2.34
Calculating CSS selectors is slow, so we need to do it after a delay.
Using the "title" attribute approach doesn't work well because, as per
browser behavior, if it is set after a timeout, then it may not show up
and user would need to move the mouse to see the native tooltip. This
change improves performance and it is more notice-able in CSS heavy
pages on slow machines.